pyproject.toml
[project]
name = "quickappoint"
version = "0.1.0"
description = ""
authors = [
    {name = "b-kostenko",email = "bogdankostenko83@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.10,<4.0"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "loguru (>=0.7.3,<0.8.0)",
    "sqlalchemy (>=2.0.41,<3.0.0)",
    "pydantic-settings (>=2.9.1,<3.0.0)",
    "alembic (>=1.16.1,<2.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic[email] (>=2.11.5,<3.0.0)",
    "passlib[bcrypt] (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "isort (>=6.0.1,<7.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
